invoice generator

+++ db/migrate/20170401220029_add_price_to_horse_activity.rb
@@ -1,5 +1,5 @@
 class AddPriceToHorseActivity < ActiveRecord::Migration[5.0]
   def change
-    add_column :horse_activities, :price, :integer
+    add_column :horse_activities, :price, :float
   end
 end
