+++ db/migrate/20170411201833_add_balance_to_owner_payment.rb
@@ -1,5 +1,5 @@
 class AddBalanceToOwnerPayment < ActiveRecord::Migration[5.0]
   def change
-    add_column :owner_payments, :balance, :integer
+    add_column :owner_payments, :balance, :float
   end
 end
